ATO-2737: Move user identity service into shared folder
