+++ ipv-stub-template.yml
@@ -98,6 +98,7 @@ Resources:
           # These will always be over-written when running locally in parameters.json
           DYNAMO_ENDPOINT: !If [IsLocalEnv, "", !Ref AWS::NoValue]
           DUMMY_JWKS: !If [IsLocalEnv, "", !Ref AWS::NoValue]
+          IDENTITY_STUB: "IpvStub"
       Events:
         Get:
           Type: Api
@@ -158,6 +159,7 @@ Resources:
           ORCH_PUBLIC_SIGNING_JWKS_URL:
             !FindInMap [EnvironmentConfiguration, !Ref Environment, orchJwksUrl]
           IPV_AUDIENCE: !Sub https://identity.${Environment}.account.gov.uk/
+          IDENTITY_STUB: "IpvStub"
       Events:
         Get:
           Type: Api
@@ -215,6 +217,7 @@ Resources:
       Environment:
         Variables:
           ENVIRONMENT: !Sub ${Environment}
+          IDENTITY_STUB: "IpvStub"
       Events:
         Get:
           Type: Api
