+++ src/main/ipv-stub/data/ipv-dummy-constants.ts
@@ -1,4 +1,4 @@
-import { UserIdentity } from "../interfaces/user-identity-interface.ts";
+import { UserIdentity } from "../../shared-identity/interfaces/user-identity-interface.ts";
 
 const env =
   process.env.ENVIRONMENT == "dev" ? "authdev3.dev" : process.env.ENVIRONMENT;
